Fix RefBase Explicit destruction
For an object to be managed by std smart pointer, its base class should NOT be "RefBase".  It is not only unnecessary but also risking the object being owned by android smart pointer as well, which may destroy the object two times.

This issue may be detected by the destructor of class RefBase, with the log complaint "RefBase: Explicit destruction,..."

Test: monkey test for one day and one night

Signed-off-by: Jintao Zhu <zhujtcsieee@gmail.com>
Change-Id: I853ca92b9176eb70b25c6a4acf1a32dce1d8f8d5
Signed-off-by: Akash Srivastava <akashniki@gmail.com>